third_party/WebKit/LayoutTests/fast/dom/Window/window-closed-immediate-expected.txt

window.closed should return true after window.close().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS w.closed is true
PASS w.closed is true
PASS successfullyParsed is true

TEST COMPLETE
